+++ src/main/java/ddobab/blank/web/QuestionApiV1Controller.java
@@ -1,6 +1,7 @@
 package ddobab.blank.web;
 
 import ddobab.blank.domain.question.QuestionCategory;
+import ddobab.blank.exception.customException.UnauthorizedException;
 import ddobab.blank.security.annotation.LoginUser;
 import ddobab.blank.security.dto.SessionUserDto;
 import ddobab.blank.service.question.QuestionService;
@@ -27,46 +28,55 @@ public class QuestionApiV1Controller {
     @PostMapping
     public ResponseEntity<ResponseDto<QuestionResponseDto>> save(@LoginUser SessionUserDto loginUser, @RequestBody QuestionSaveRequestDto requestDto) {
         //bean validation 필요
-        QuestionResponseDto data = questionService.save(loginUser.getNo(), requestDto);
-        //log 필요
-        return new ResponseEntity<>(new ResponseDto<>(data, null), HttpStatus.CREATED);
+        if(loginUser!=null){
+            QuestionResponseDto data = questionService.save(loginUser.getNo(), requestDto);
+            return new ResponseEntity<>(new ResponseDto<>(data, null), HttpStatus.CREATED);
+        }else{
+            throw new UnauthorizedException("데이터 생성 권한이 없습니다.");
+        }
     }
 
     @GetMapping("/{no}")
     public ResponseEntity<ResponseDto<QuestionResponseDto>> getQuestion(@PathVariable Long no) {
-        //pathvariable type mismatch? ex
         QuestionResponseDto data = questionService.findByNo(no);
-
         return new ResponseEntity<>(new ResponseDto<>(data, null), HttpStatus.OK);
     }
 
     @PutMapping("/{no}")
     public ResponseEntity<ResponseDto<QuestionResponseDto>> update(@LoginUser SessionUserDto loginUser, @PathVariable Long no, @RequestBody QuestionUpdateRequestDto requestDto) {
         //bean valid
-        Long writerNo = questionService.getQuestionWrtier(no);
-        if(writerNo.equals(loginUser.getNo())) {
-            QuestionResponseDto data = questionService.update(no, requestDto);
-            return new ResponseEntity<>(new ResponseDto<>(data, null), HttpStatus.ACCEPTED);
-        } else {
-            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);//예외처리, 메세지 넘기기
+        if(loginUser!=null){
+            Long writerNo = questionService.getQuestionWriter(no);
+            if(writerNo.equals(loginUser.getNo())) {
+                QuestionResponseDto data = questionService.update(no, requestDto);
+                return new ResponseEntity<>(new ResponseDto<>(data, null), HttpStatus.ACCEPTED);
+            } else {
+                throw new UnauthorizedException("데이터 변경 권한이 없습니다.");
+            }
+        }else{
+            throw new UnauthorizedException("데이터 변경 권한이 없습니다.");
         }
+
     }
 
     @DeleteMapping("/{no}")
     public ResponseEntity<ResponseDto<?>> delete(@LoginUser SessionUserDto loginUser, @PathVariable Long no) {
-        //path var ex
-        Long writerNo = questionService.getQuestionWrtier(no);
-        if (writerNo.equals(loginUser.getNo())) {
-            questionService.delete(no);
-            return new ResponseEntity<>(new ResponseDto<>(null, null),HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);   //예외처리 메세지
+        if(loginUser!=null){
+            Long writerNo = questionService.getQuestionWriter(no);
+            if (writerNo.equals(loginUser.getNo())) {
+                questionService.delete(no);
+                return new ResponseEntity<>(new ResponseDto<>(null, null),HttpStatus.OK);
+            } else {
+                throw new UnauthorizedException("데이터 삭제 권한이 없습니다.");
+            }
+        }else{
+            throw new UnauthorizedException("데이터 삭제 권한이 없습니다.");
         }
+
     }
 
     @GetMapping("/category")
     public ResponseEntity<ResponseDto<QuestionCategory[]>> getCategoryList() {
-
         QuestionCategory[] data = QuestionCategory.values();
         return new ResponseEntity<>(new ResponseDto<>(data, null), HttpStatus.OK);
     }
